requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
transformers
torch
gdown
